super, this, polimorfismo de metodos, sobreescribiendo metodom tosString y interfaces
