+++ src/Patient.java
@@ -1,11 +1,5 @@
-public class Patient {
-
-    //Atributos
-    int id;
-    private String name;
-    private String email;
-    private String address;
-    private String phoneNumber;
+public class Patient extends User{
+    //Atributo
     private String birthday;
     private double weight;
     private double height;
@@ -13,8 +7,7 @@ public class Patient {
 
     //Constructor
     Patient(String name, String email){
-        this.name = name;
-        this.email = email;
+        super(name, email);
     }
 
     public void setWeight( double weight ){
@@ -25,48 +18,9 @@ public class Patient {
       return this.weight + " Kg.";
     }
 
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
 
     public void setPhoneNumber(String phoneNumber) {
-        if (phoneNumber.length() > 8 ){
-            System.out.println("El numero telefonico debe ser de 8 digitos maximo");
-        }else if(phoneNumber.length() == 8){
-            this.phoneNumber = phoneNumber;
-        }
+
     }
 
     public String getBirthday() {
@@ -91,5 +45,10 @@ public class Patient {
 
     public void setBlood(String blood) {
         this.blood = blood;
+    }
+
+    @Override
+    public String toString() {
+        return super.toString() + "\nAge: "+birthday + "\n Weight: "+getWeight() +"\n Height" + getHeight() + "\nBlood: "+blood;
     }
 }
